+++ post-build.js
@@ -2,5 +2,5 @@ const fs = require("fs");
 fs.copyFile(
     "./platforms/android/app/build/outputs/apk/debug/app-debug.apk",
     "./emoji-story-cubes_debug.apk",
-    err => console.error(err)
+    err => err ? console.error(err) : console.log('copied')
 );